feat: redesign special collections page with enhanced UI and performance
- Restructured layout with video banner, horizontal category scroll, and improved product grid
- Added performance optimizations including memoization, Suspense, and custom transition hooks
- Enhanced accessibility with ARIA labels, semantic HTML, and error boundary wrapper
