+++ .gitignore
@@ -6,4 +6,4 @@ node_modules
 /.mf
 .env
 .shopify
-/.react-router
+.react-router
