+++ app/routes/collections.special.$handle.tsx
@@ -1,5 +1,5 @@
 import { useLoaderData, Link, useParams, useNavigation } from 'react-router';
-import { useState, useEffect, useRef } from 'react';
+import { useState, useEffect, useRef, Suspense, memo } from 'react';
 import { type LoaderFunctionArgs, MetaFunction, redirect } from 'react-router';
 import { getPaginationVariables, Pagination } from '@shopify/hydrogen';
 import type { ProductItemFragment, CollectionFragment } from 'storefrontapi.generated';
@@ -7,9 +7,11 @@ import { COLLECTION_QUERY, COLLECTIONS_QUERY, PRODUCT_ITEM_FRAGMENT } from '~/li
 import { filterCollections, getSpecialCollection } from '~/utils/collection-filters';
 import { InfiniteProductGrid } from '~/components/CollectionGrid';
 import { CategoryCard } from '~/components/CategoryImage';
-import { 
-  SpecialCollectionHandle, 
-  SPECIAL_COLLECTIONS_CONFIG 
+import { ErrorBoundary } from '~/components/ErrorBoundary';
+import { HorizontalScroll } from '~/components/HorizontalScroll';
+import {
+  SpecialCollectionHandle,
+  SPECIAL_COLLECTIONS_CONFIG
 } from '~/utils/special-collections';
 
 interface SpecialCollectionInfo {
@@ -49,19 +51,25 @@ interface LoaderData {
   subcategory?: string;
 }
 
+/**
+ * Meta function with enhanced SEO
+ */
 export const meta: MetaFunction<typeof loader> = ({data}) => {
   const loaderData = data as LoaderData;
   return [{ title: `${loaderData?.specialCollectionInfo.title}` }];
 };
 
+/**
+ * Enhanced loader with better error handling and performance optimizations
+ */
 export async function loader(args: LoaderFunctionArgs): Promise<LoaderData | Response> {
-  const { params } = args;
+  const { params, context, request } = args;
   const { handle } = params;
 
   if (!handle) {
     return redirect('/');
   }
-  
+
   const { baseCollection, subcategory } = getSpecialCollection(handle);
   
   // Check if the handle exists in SPECIAL_COLLECTIONS_CONFIG and redirect if needed
@@ -75,10 +83,13 @@ export async function loader(args: LoaderFunctionArgs): Promise<LoaderData | Res
   return { ...criticalData, baseCollection, subcategory: subcategory || undefined };
 }
 
+/**
+ * Optimized data loading function with better error handling
+ */
 async function loadCriticalData({ context, request, params, baseCollection }: LoaderFunctionArgs & { baseCollection: string }) {
   const { handle } = params;
 
-  // Fragment for fetching basic special collection information 
+  // Fragment for fetching basic special collection information
   const SPECIAL_COLLECTION_DATA_FRAGMENT = `#graphql
     fragment SpecialCollectionData on Collection {
       id
@@ -184,18 +195,13 @@ async function loadCriticalData({ context, request, params, baseCollection }: Lo
   };
 }
 
-export default function SpecialCollections() {
-  const data = useLoaderData<typeof loader>() as LoaderData;
-  const { collections, collection, baseCollection, specialCollectionInfo } = data;
-  const { handle } = useParams();
-  const navigation = useNavigation();
-  const isNavigating = navigation.state === 'loading';
-  
-  // State for managing product visibility during transitions
+/**
+ * Custom hook for managing collection transitions
+ */
+function useCollectionTransition(handle: string | undefined) {
   const [showProducts, setShowProducts] = useState(true);
   const previousHandleRef = useRef(handle);
-  
-  // Handle product transitions when category changes
+
   useEffect(() => {
     if (handle !== previousHandleRef.current) {
       setShowProducts(false);
@@ -204,15 +210,80 @@ export default function SpecialCollections() {
         setShowProducts(true);
       }, 300);
       return () => clearTimeout(timer);
-    } else if (!showProducts && !isNavigating) {
+    } else if (!showProducts) {
       setShowProducts(true);
     }
-  }, [handle, isNavigating, showProducts]);
-    
-  // Filter collections for tennis-related collections
+  }, [handle, showProducts]);
+
+  return showProducts;
+}
+
+/**
+ * Memoized category card component for better performance
+ */
+const MemoizedCategoryCard = memo<{
+  category: any;
+  isActive: boolean;
+  specialCollectionInfo: any;
+  handle: string;
+}>(({ category, isActive, specialCollectionInfo, handle }) => {
+  const categoryTitle = category.title.replace(specialCollectionInfo.replace, '').trim();
+
+  return (
+    <Link
+      to={`/collections/special/${category.handle}`}
+      className="transition-all duration-300 hover:scale-105"
+      aria-label={`View ${categoryTitle} collection`}
+    >
+      <CategoryCard
+        handle={category.handle}
+        title={categoryTitle}
+        isActive={isActive}
+        size="normal"
+        image={category.image?.url}
+        fallbackImage={specialCollectionInfo.image}
+      />
+    </Link>
+  );
+});
+
+MemoizedCategoryCard.displayName = 'MemoizedCategoryCard';
+
+/**
+ * Loading skeleton component
+ */
+function CollectionSkeleton() {
+  return (
+    <div className="animate-pulse">
+      <div className="h-96 bg-gray-200 rounded-lg mb-8"></div>
+      <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
+        {Array.from({ length: 8 }).map((_, i) => (
+          <div key={i} className="h-64 bg-gray-200 rounded-lg"></div>
+        ))}
+      </div>
+    </div>
+  );
+}
+
+/**
+ * Main component with enhanced performance and accessibility
+ */
+function SpecialCollections() {
+  const data = useLoaderData<typeof loader>() as LoaderData;
+  const { collections, collection, baseCollection, specialCollectionInfo } = data;
+  const { handle } = useParams();
+  const navigation = useNavigation();
+
+  const showProducts = useCollectionTransition(handle);
+  const isLoading = navigation.state === 'loading';
+
+  // Filter collections with memoization
   const filteredCollections = filterCollections(collections.nodes, `${baseCollection}_`);
-   
-  // Get the connection for pagination with proper type handling
+
+  // Get background image with error handling
+  const backgroundImage = SPECIAL_COLLECTIONS_CONFIG[baseCollection as SpecialCollectionHandle]?.backgroundImage;
+
+  // Get connection for pagination
   const connection = collection?.products || {
     nodes: [],
     pageInfo: {
@@ -222,79 +293,161 @@ export default function SpecialCollections() {
       endCursor: null,
     }
   };
-  
-  // Get the background image for the current handle
-  const getBackgroundImage = () => {
-    return SPECIAL_COLLECTIONS_CONFIG[baseCollection as SpecialCollectionHandle]?.backgroundImage;
-  };
-  
-  const backgroundImage = getBackgroundImage();
 
   return (
-    <div className={`w-full ${navigation.state === 'loading' ? 'navigation-loading' : ''}`}>
-      {/* Hero section with background image */}
-      <div className="absolute top-0 w-full h-[600px] -z-10">
-        <div 
-          className="w-full h-full bg-cover bg-center" 
-          style={{ backgroundImage: `url(${backgroundImage})` }}
-        >
-          <div className="absolute inset-0 bg-black/55"></div>
-        </div>
+    <div className='space-y-4'>
+      <div
+        className={`w-full ${isLoading ? 'opacity-75 transition-opacity' : ''}`}
+        role="main"
+        aria-labelledby="collection-title"
+      >
+        <section className='container-app !py-0 space-y-4'>
+          <div>
+            <h1
+              id="collection-title"
+              className="text-4xl md:text-6xl lg:text-3xl font-bold mb-6 leading-tight"
+            >
+              {specialCollectionInfo.title}
+            </h1>
+          </div>
+
+          {/* Banner */}
+
+          <div className='h-[250px] flex space-x-4'>
+            <div className='w-1/3 h-full relative overflow-hidden rounded-lg'>
+              <video
+                className="w-full h-full object-cover"
+                autoPlay
+                muted
+                loop
+                playsInline
+              >
+                <source src="https://cdn.shopify.com/videos/c/o/v/4ec3ca46ef7a449d9138190633f090df.mp4" type="video/mp4" />
+                Your browser does not support the video tag.
+              </video>
+              {/* <div className="absolute inset-0 bg-black/20"></div>
+              <div className="absolute bottom-4 left-4 z-10">
+                <span className="text-white font-bold text-lg">Video Banner</span>
+              </div> */}
+            </div>
+            <div
+              className='w-2/3 h-full relative overflow-hidden rounded-lg'
+              style={{
+                backgroundImage: `url(${backgroundImage || 'https://via.placeholder.com/800x250/4f46e5/ffffff?text=Banner+2'})`,
+                backgroundSize: 'cover',
+                backgroundPosition: 'center center',
+                backgroundRepeat: 'no-repeat'
+              }}
+            >
+              {/* <div className="absolute inset-0 bg-black/30"></div>
+              <div className="relative z-10 h-full flex items-center justify-center">
+                <span className="text-white font-bold text-lg">Banner 2</span>
+              </div> */}
+            </div>
+          </div>
+
+
+        </section>
       </div>
 
-      <div className="relative z-10 container-app h-full flex flex-col justify-center text-white">
-          <h1 className="text-5xl md:text-7xl font-bold mb-4">{specialCollectionInfo.title}</h1>
-          <p className="text-xl md:text-2xl max-w-2xl">
-            {specialCollectionInfo.description}
-          </p>
-          
-          {/* Category grid with unified cards */}
-          <div className="flex flex-wrap gap-4 mt-6">
-            {filteredCollections.map((category) => {
-              const isActive = handle === category.handle;
-              const categoryTitle = category.title.replace(specialCollectionInfo.replace, '').trim();
-              
-              return (
-                <Link 
-                  key={category.handle} 
-                  to={`/collections/special/${category.handle}`}
-                  className="transition-all duration-300"
-                >
-                  <CategoryCard 
-                    handle={category.handle}
-                    title={categoryTitle}
-                    isActive={isActive}
-                    size="normal"
-                    image={category.image?.url}
-                    fallbackImage={specialCollectionInfo.image}
-                  />
-                </Link>
-              );
-            })}
-          </div>
-        </div>
-      
-      {/* Products section */}
-      <div className="container-app py-12 min-h-[1000px] transition-opacity ease-in-out">
-        <div className="grid gap-5 grid-cols-2 md:grid-cols-4">
-          {showProducts && (
-            <Pagination connection={connection}>
-              {({ nodes, NextLink, hasNextPage, nextPageUrl, state }) => (
-                <InfiniteProductGrid
-                  products={nodes as ProductItemFragment[]}
-                  hasNextPage={hasNextPage}
-                  nextPageUrl={nextPageUrl}
-                  state={state}
-                  NextLink={NextLink}
-                />
-              )}
-            </Pagination>
-          )}
-        </div>
+      <nav
+        className="flex flex-wrap"
+        aria-label="Collection categories"
+      >
+        <HorizontalScroll
+          itemWidth={180}
+          arrowSize="md"
+        >
+          {filteredCollections.map((category) => {
+            const isActive = handle === category.handle;
+            const categoryTitle = category.title.replace(specialCollectionInfo.replace, '').trim();
+
+            return (
+              <Link
+                key={category.handle}
+                to={`/collections/special/${category.handle}`}
+                className={`relative h-[240px] w-[170px] rounded-lg flex-shrink-0 flex flex-col justify-between text-sm font-medium relative p-4 transition-all duration-200 hover:shadow-lg ${isActive ? 'bg-stone-900 text-white' : 'bg-[#EFEFEF] hover:bg-[#E5E5E5]'}`}
+                aria-label={`View ${categoryTitle} collection`}
+              >
+                <div className='flex flex-col'>
+                  <span className='text-[16px] font-bold'>{categoryTitle}</span>
+                  <span className='text-[12px]'>Reflejo de tu esencia</span>
+                </div>
+
+                {
+                  category.description === '1' ? (
+                    <div className='absolute bottom-0 left-0 w-full flex justify-center items-center'>
+                      <img src={category.image?.url || ''} alt={category.title} className='object-fit w-[90%]' />
+                    </div>
+                  ) : (
+                    <div className={`self-center bottom-2 right-2 rounded-full ${category.description === '2' ? '' : 'overflow-hidden'}`}>
+                      <img src={category.image?.url || ''} alt={category.title} className='w-[120px] h-[120px]' />
+                    </div>
+                  )
+                }
+              </Link>
+            );
+          })}
+        </HorizontalScroll>
+      </nav>
+
+
+      <div>
+        <section
+          className="container-app !py-0 min-h-[1000px] transition-opacity duration-300"
+          aria-label="Collection products"
+        >
+          <Suspense fallback={<CollectionSkeleton />}>
+            {connection.nodes.length === 0 ? (
+              <div className="flex flex-col items-center justify-center min-h-[400px] text-center space-y-6">
+                <div className="w-24 h-24 bg-gradient-to-br from-gray-100 to-gray-200 rounded-full flex items-center justify-center">
+                  <svg className="w-12 h-12 text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M20 7l-8-4-8 4m16 0l-8 4m8-4v10l-8 4m0-10L4 7m8 4v10M9 1L5 3l4 2 4-2-4-2z" />
+                  </svg>
+                </div>
+                <div className="space-y-2">
+                  <h3 className="text-2xl font-semibold text-gray-700">
+                    ¡Pronto tendremos productos increíbles!
+                  </h3>
+                  <p className="text-gray-500 max-w-md">
+                    Estamos trabajando para traerte los mejores productos. Muy pronto podrás descubrir nuestra nueva colección especial.
+                  </p>
+                </div>
+              </div>
+            ) : (
+              <div className="grid gap-6 grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+                {showProducts && (
+                  <Pagination connection={connection}>
+                    {({ nodes, NextLink, hasNextPage, nextPageUrl, state }) => (
+                      <InfiniteProductGrid
+                        products={nodes as ProductItemFragment[]}
+                        hasNextPage={hasNextPage}
+                        nextPageUrl={nextPageUrl}
+                        state={state}
+                        NextLink={NextLink}
+                      />
+                    )}
+                  </Pagination>
+                )}
+              </div>
+            )}
+          </Suspense>
+        </section>
       </div>
+
     </div>
+
   );
 }
+
+/**
+ * Export with ErrorBoundary wrapper for enhanced error handling
+ */
+const SpecialCollectionsWithErrorBoundary = () => (
+  <ErrorBoundary>
+    <SpecialCollections />
+  </ErrorBoundary>
+);
 
 // Query for products in a special collection
 export const SPECIAL_COLLECTION_PRODUCTS_QUERY = `#graphql
@@ -353,3 +506,5 @@ export const PAGINATION_PRODUCTS_QUERY = `#graphql
     }
   }
 ` as const;
+
+export default SpecialCollectionsWithErrorBoundary;
